+++ examples/static-site-generator/playground-export-smoke.mjs
@@ -280,6 +280,14 @@ $amp_id = wp_insert_post(array(
 	'post_content' => '<p>AMP companion export target.</p>',
 ));
 
+$preloaded_document_id = wp_insert_post(array(
+	'post_type' => 'page',
+	'post_status' => 'publish',
+	'post_title' => 'Preloaded Document',
+	'post_name' => 'preloaded-document',
+	'post_content' => '<p>Document preload export target.</p>',
+));
+
 wp_update_post(array(
 	'ID' => $child_id,
 	'post_parent' => $parent_id,
@@ -302,6 +310,7 @@ $svg_link_url = get_permalink($svg_link_id);
 $image_metadata_url = get_permalink($image_metadata_id);
 $schema_profile_url = get_permalink($schema_profile_id);
 $amp_url = get_permalink($amp_id);
+$preloaded_document_url = get_permalink($preloaded_document_id);
 $protocol_child_url = preg_replace('/^https?:/', '', $child_url);
 $rest_route_url = '/?rest_route=/wp/v2/posts';
 $feed_query_url = '/?feed=rss2';
@@ -354,6 +363,7 @@ $static_content = '<p id="section">Static smoke page.</p>'
 	. '<svg><a xlink:href="' . esc_url($svg_link_url) . '"><text>SVG link</text></a></svg>'
 	. '<link rel="amphtml" href="' . esc_url($amp_url) . '">'
 	. '<link rel="manifest" href="' . esc_url($manifest_url) . '">'
+	. '<link rel="preload" as="document" href="' . esc_url($preloaded_document_url) . '">'
 	. '<link rel="preload" as="fetch" href="' . esc_url($player_config_url) . '">'
 	. '<link rel="preload" as="image" href="' . esc_url($asset_url) . '" imagesrcset="' . esc_url($asset_url) . ' 1x, ' . esc_url($asset_url . '?preload=2x') . ' 2x">'
 	. '<video><track kind="captions" src="' . esc_url($captions_url . '?track=1') . '"></video>'
@@ -454,6 +464,7 @@ $scoped_svg_link_url = get_permalink($svg_link_id);
 $scoped_image_metadata_url = get_permalink($image_metadata_id);
 $scoped_schema_profile_url = get_permalink($schema_profile_id);
 $scoped_amp_url = get_permalink($amp_id);
+$scoped_preloaded_document_url = get_permalink($preloaded_document_id);
 $scoped_protocol_child_url = preg_replace('/^https?:/', '', $scoped_child_url);
 $scoped_rest_route_url = home_url('/?rest_route=/wp/v2/posts');
 $scoped_feed_query_url = home_url('/?feed=rss2');
@@ -514,6 +525,7 @@ $scoped_static_content = '<p id="section">Static smoke page.</p>'
 	. '<svg><a xlink:href="' . esc_url($scoped_svg_link_url) . '"><text>SVG link</text></a></svg>'
 	. '<link rel="amphtml" href="' . esc_url($scoped_amp_url) . '">'
 	. '<link rel="manifest" href="' . esc_url($scoped_manifest_url) . '">'
+	. '<link rel="preload" as="document" href="' . esc_url($scoped_preloaded_document_url) . '">'
 	. '<link rel="preload" as="fetch" href="' . esc_url($scoped_player_config_url) . '">'
 	. '<link rel="preload" as="image" href="' . esc_url($scoped_asset_url) . '" imagesrcset="' . esc_url($scoped_asset_url) . ' 1x, ' . esc_url($scoped_asset_url . '?preload=2x') . ' 2x">'
 	. '<p><a class="other-scope-link" href="https://playground.wordpress.net/scope:other-site/static-page/">Other scope</a></p>'
@@ -656,6 +668,7 @@ async function verifyExport() {
 	assertFile('image-metadata/index.html');
 	assertFile('schema-profile/index.html');
 	assertFile('amp-companion/index.html');
+	assertFile('preloaded-document/index.html');
 	assertFile('parent-page/index.html');
 	assertFile('parent-page/child-page/index.html');
 	assertFile('wp-content/uploads/ssgwp-smoke-asset.txt');
@@ -694,6 +707,7 @@ async function verifyExport() {
 		'../image-metadata/index.html',
 		'../schema-profile/index.html',
 		'../amp-companion/index.html',
+		'../preloaded-document/index.html',
 		'relative-child/index.html',
 		'index.html#section',
 		'../wp-content/uploads/ssgwp-smoke-asset.txt?meta=1',
@@ -970,6 +984,7 @@ async function verifyScopedExport() {
 	assertFile('image-metadata/index.html');
 	assertFile('schema-profile/index.html');
 	assertFile('amp-companion/index.html');
+	assertFile('preloaded-document/index.html');
 	assertFile('parent-page/child-page/index.html');
 	assertFile('wp-content/uploads/ssgwp-smoke-asset.txt');
 	assertFile('wp-content/uploads/ssgwp-smoke-captions.vtt');
@@ -1020,6 +1035,7 @@ async function verifyScopedExport() {
 		'../image-metadata/index.html',
 		'../schema-profile/index.html',
 		'../amp-companion/index.html',
+		'../preloaded-document/index.html',
 		'relative-child/index.html',
 		'index.html#section',
 		'../wp-content/uploads/ssgwp-smoke-asset.txt?meta=1',